backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
duckdb>=1.0.0
pandas>=2.1.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
